pyproject.toml
[project]
name = "advent-bot-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=24.10.0",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.7",
    "requests>=2.32.3",
    "spotipy>=2.24.0",
]
